Logout Funcitionality
Not fully completed, but a bare bones version. So once the user logouts their refresh token will be blacklisted so it cannot be used again. Added a logged in state where if the user is logged in they will only have the option to log out or if the user is not logged in they will only have the option to sign up or login. Also added axios interceptors, so that when an access token is expired it knows to grab a new one using the refresh token. Most of the code is just error checking. This is required because most of the time the user will be on the website because they are studying.