Edit Room Functionality
Improved semantics, now there are now separate files for the room info card and edit room form. This is more efficient and a better style of code. Each room info card will handle its own edit form modal, which fills the form with the current info. Also, implemented a state that handles whether a room has been updated. This allows for in page re-rendering of all the users rooms rather than having to refresh the page.