Fix to specific user Spotify Token creations
Previously when a user would log into their spotify account it would not generate a new spotify token. This was because I was creating the token under request.user in the spotify callback function which would be the django superuser. To fix this I used the axiosInstance containing the user logged in to make the initial call to get spotify auth url. I stored the id of the user sending this request in session so that when they login to their spotify account when they are redirected to the uri by spotify they can retrieve the initial user from the session.